feature: atualização do package e corrigido ao alterUserDeleUsername para realizar a migration

+++ src/database/migrations/1692659742047-AlterUserDeleteUsername.ts
@@ -1,19 +1,27 @@
-import { MigrationInterface, QueryRunner, TableColumn } from "typeorm"
+import { MigrationInterface, QueryRunner, TableColumn } from "typeorm";
 
 export class AlterUserDeleteUsername1692659742047 implements MigrationInterface {
-
     public async up(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.dropColumn("users", "username");
+        const table = await queryRunner.getTable("users");
+        const usernameColumn = table?.findColumnByName("username");
+
+        if (usernameColumn) {
+            await queryRunner.dropColumn("users", "username");
+        }
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.addColumn(
-            "users",
-            new TableColumn({
-                name: "username",
-                type: "varchar"
-            })
-        )
-    }
+        const table = await queryRunner.getTable("users");
+        const usernameColumn = table?.findColumnByName("username");
 
+        if (!usernameColumn) {
+            await queryRunner.addColumn(
+                "users",
+                new TableColumn({
+                    name: "username",
+                    type: "varchar",
+                }),
+            );
+        }
+    }
 }
